test: first test for GatheringPlacesRepository, fixed CREATE queries; added test library dependencies in go.sum and go.mod
