+++ app/entities/meeting/query/FindByID.go
@@ -7,7 +7,7 @@ import (
 )
 
 func FindByID(ctx context.Context, id uuid.UUID, db *sql.DB) *sql.Row {
-	const query = `SELECT FROM meetings WHERE id = $1`
+	const query = `SELECT * FROM meetings WHERE id = $1`
 	row := db.QueryRow(query, id)
 	return row
 }
