+++ app/entities/user/query/FindUserByID.go
@@ -7,7 +7,7 @@ import (
 )
 
 func FindUserByID(ctx context.Context, id uuid.UUID, db *sql.DB) *sql.Row {
-	const query = `SELECT FROM users WHERE id = $1`
+	const query = `SELECT id, username, displayname, rating, current_meeting_id FROM users WHERE id = $1`
 	row := db.QueryRow(query, id)
 	return row
 }
